# Optimize log statement in ingest path