@replay 2019-07-11
actually the benchmark was only for myself to verify that my change improves the situation. now that the change itself has already been merged via another PR, we might as well just close this one