@Dieterbe 2019-07-11
no, i think that concern doesn't belong in benchmarks in the input package as explained in the comment on the code.
but anyway, i'm fine with just closing this too.